+++ Index.html
@@ -9,8 +9,8 @@
 </head>
 <body>
     
-    <div class="start-game-field">
-        <div class="header"></div>
+<div class="start-game-field">
+        
     <div class="start-box start">
         <p class="start-box--text"> Выбери сложность</p>
         <div class="start-box--difficult">
@@ -26,10 +26,10 @@
             </div>
         </div>
         <button class="start-box--button">Старт</button>
-
-        
     </div>
+
 </div>
+
 <script type="module" src="./main.js"></script>
 </body>
